+++ src/features/profile/queries/useFeedbackMutation.ts
@@ -1,31 +1,16 @@
-import { useNavigation } from '@react-navigation/native'
 import { useMutation } from 'react-query'
 
 import { api } from 'api/api'
 import { PostFeedbackBody } from 'api/gen'
-import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
-import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 
-export const useFeedbackMutation = () => {
-  // TODO(PC-35699): Move onSuccess and onError callback functions out of this mutation
-  const { navigate } = useNavigation<UseNavigationType>()
-  const navigateToProfile = () => navigate(...getTabNavConfig('Profile'))
-  const { showSuccessSnackBar, showErrorSnackBar } = useSnackBarContext()
+type FeedbackMutationOptions = {
+  onSuccess?: () => void
+  onError?: (error: unknown) => void
+}
 
+export const useFeedbackMutation = ({ onSuccess, onError }: FeedbackMutationOptions) => {
   return useMutation((body: PostFeedbackBody) => api.postNativeV1Feedback(body), {
-    onSuccess: () => {
-      showSuccessSnackBar({
-        message: 'Ta suggestion a bien été transmise\u00a0!',
-        timeout: SNACK_BAR_TIME_OUT,
-      })
-      navigateToProfile()
-    },
-    onError: () => {
-      showErrorSnackBar({
-        message: 'Une erreur s’est produite lors de l’envoi de ta suggestion. Réessaie plus tard.',
-        timeout: SNACK_BAR_TIME_OUT,
-      })
-    },
+    onSuccess,
+    onError,
   })
 }
